# backend/openapi/credence_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Credence AI",
    "version": "1.0.0"
  },
  "paths": {
...
    "/oracle/{wallet}": {
      "get": {
        "tags": [
          "Oracle Registry"
        ],
        "summary": "Get Oracle Attestations By Wallet",
        "description": "Retrieves all published oracle attestations for a given wallet.",
        "operationId": "get_oracle_attestations_by_wallet_oracle__wallet__get",
        "parameters": [
          {
            "name": "wallet",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Wallet"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OracleAttestation"
                  },
                  "title": "Response Get Oracle Attestations By Wallet Oracle  Wallet  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/governance/audit": {
      "get": {
        "tags": [
          "Governance & Administration"
        ],
        "summary": "Get Audit Timeline",
        "operationId": "get_audit_timeline_governance_audit_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Keyword logs search query",
              "default": "",
              "title": "Q"
            },
            "description": "Keyword logs search query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditLog"
                  },
                  "title": "Response Get Audit Timeline Governance Audit Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/governance/oracle": {
      "post": {
        "tags": [
          "Governance & Administration"
        ],
        "summary": "Register Oracle",
        "operationId": "register_oracle_governance_oracle_post",
        "parameters": [
          {
            "name": "x-actor",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/governance/oracle/{id}": {
      "patch": {
        "tags": [
          "Governance & Administration"
        ],
        "summary": "Update Oracle Status",
        "operationId": "update_oracle_status_governance_oracle__id__patch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "x-actor",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/governance/policy": {
      "post": {
        "tags": [
          "Governance & Administration"
        ],
        "summary": "Register Policy Rule",
        "operationId": "register_policy_rule_governance_policy_post",
        "parameters": [
          {
            "name": "x-actor",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/governance/passport/{id}": {
      "patch": {
        "tags": [
          "Governance & Administration"
        ],
        "summary": "Update Passport State",
        "operationId": "update_passport_state_governance_passport__id__patch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "x-actor",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/governance/proposals": {
      "post": {
        "tags": [
          "Governance & Administration"
        ],
        "summary": "Create Gov Proposal",
        "operationId": "create_gov_proposal_governance_proposals_post",
        "parameters": [
          {
            "name": "x-actor",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Proposal"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Governance & Administration"
        ],
        "summary": "List Gov Proposals",
        "operationId": "list_gov_proposals_governance_proposals_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Proposal"
                  },
                  "title": "Response List Gov Proposals Governance Proposals Get"
                }
              }
            }
          }
        }
      }
    },
    "/governance/proposals/{id}/approve": {
      "post": {
        "tags": [
          "Governance & Administration"
        ],
        "summary": "Approve Gov Proposal",
        "operationId": "approve_gov_proposal_governance_proposals__id__approve_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "x-actor",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/governance/proposals/{id}/execute": {
      "post": {
        "tags": [
          "Governance & Administration"
        ],
        "summary": "Execute Gov Proposal",
        "operationId": "execute_gov_proposal_governance_proposals__id__execute_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "x-actor",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuditLog": {
        "properties": {
          "log_id": {
            "type": "string",
            "title": "Log Id"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "performed_by": {
            "type": "string",
            "title": "Performed By"
          },
          "resource": {
            "type": "string",
            "title": "Resource"
          },
          "result": {
            "type": "string",
            "title": "Result"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "log_id",
          "action",
          "performed_by",
          "resource",
          "result",
          "timestamp"
        ],
        "title": "AuditLog"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OracleAttestation": {
        "properties": {
          "attestation_id": {
            "type": "string",
            "title": "Attestation Id"
          },
          "wallet": {
            "type": "string",
            "title": "Wallet"
          },
          "offer_id": {
            "type": "string",
            "title": "Offer Id"
          },
          "offer_hash": {
            "type": "string",
            "title": "Offer Hash"
          },
          "attestation_hash": {
            "type": "string",
            "title": "Attestation Hash"
          },
          "oracle_address": {
            "type": "string",
            "title": "Oracle Address"
          },
          "credit_score": {
            "type": "integer",
            "title": "Credit Score"
          },
          "risk_level": {
            "type": "string",
            "title": "Risk Level"
          },
          "issued_at": {
            "type": "string",
            "format": "date-time",
            "title": "Issued At"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "signature": {
            "type": "string",
            "title": "Signature"
          },
          "verified": {
            "type": "boolean",
            "title": "Verified"
          },
          "oracle_version": {
            "type": "string",
            "title": "Oracle Version"
          },
          "chain_id": {
            "type": "integer",
            "title": "Chain Id"
          }
        },
        "type": "object",
        "required": [
          "attestation_id",
          "wallet",
          "offer_id",
          "offer_hash",
          "attestation_hash",
          "oracle_address",
          "credit_score",
          "risk_level",
          "issued_at",
          "expires_at",
          "signature",
          "verified",
          "oracle_version",
          "chain_id"
        ],
        "title": "OracleAttestation"
      },
...
      "Proposal": {
        "properties": {
          "proposal_id": {
            "type": "string",
            "title": "Proposal Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "submitted_by": {
            "type": "string",
            "title": "Submitted By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "proposal_id",
          "title",
          "type",
          "status",
          "submitted_by",
          "created_at"
        ],
        "title": "Proposal"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}